Move map::tests to its own file
